+++ gulpfile.js
@@ -55,7 +55,8 @@ var paths = {
   curatedTypings: [
     'src/*.d.ts',
     '!src/instance-id.d.ts',
-    '!src/project-management.d.ts'
+    '!src/project-management.d.ts',
+    '!src/messaging.d.ts',
   ],
 };
 
@@ -68,7 +69,6 @@ const TEMPORARY_TYPING_EXCLUDES = [
   '!lib/database/*.d.ts',
   '!lib/firestore/*.d.ts',
   '!lib/machine-learning/*.d.ts',
-  '!lib/messaging/*.d.ts',
   '!lib/remote-config/*.d.ts',
   '!lib/security-rules/*.d.ts',
   '!lib/storage/*.d.ts',
